+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## v1.21.0
+
+### Changed
+
+- CloudFrontBucket and CloudFrontBucketV2 are deprecated in favor of ExtendedBucket.
+- ExtendedBucket now supports creating origins for CloudFront.
+- The deploy method on \*Bucket types no longer accepts a scope.
+
+### Removed
+
+### Added
+
 ## v1.20.0
 
 ### Changed